Collapse navbar and header into single component
