--- addon/styles/components/es-navbar.css
@@ -1,181 +1,0 @@
-.es-navbar {
-  background-color: #1C1E23;
-  color: white;
-  display: flex;
-  flex-direction: column;
-  /* font-size: var(--text-preset-2); */
-  font-size: 1.2em;
-  font-weight: 500;
-  max-width: var(--container-width-large);
-  padding: 0.5rem;
-  width: 100%;
-}
-
-.navbar-brand-wrapper {
-  padding: 1rem;
-  /* TODO using order is not ideal for accessibility */
-  order: 2;
-}
-
-.navbar-brand {
-  height: 5rem;
-}
-
-.navbar-list {
-  padding: 0;
-  width: 100%;
-  margin: 0.5rem 0 0 0;
-  /* TODO using order is not ideal for accessibility */
-  order: 3;
-  display: none;
-}
-
-.navbar-list[aria-expanded="true"] {
-  display: inline-block;
-}
-
-.navbar-toggler {
-  background-color: transparent;
-  border: 1px solid transparent;
-  border-radius: 4px;
-  border-color: var(--color-light);
-  margin-left: auto;
-  /* TODO using order is not ideal for accessibility */
-  order: 1;
-  margin-bottom: -3rem;
-  height: 3rem;
-}
-
-.navbar-toggler .navbar-toggler-icon {
-  background-image: url("data:image/svg+xml;charset=utf8,%3Csvg viewBox='0 0 30 30' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath stroke='%23ffffff' stroke-width='2' stroke-linecap='round' stroke-miterlimit='10' d='M4 7h22M4 15h22M4 23h22'/%3E%3C/svg%3E")
-}
-
-.navbar-toggler-icon {
-  background-size: 100% 100%;
-  background: center center no-repeat;
-  content: "";
-  display: block;
-  height: 1.5rem;
-  width: 2.5rem;
-}
-
-.navbar-list-item {
-  list-style: none;
-  position: relative;
-}
-
-.navbar-list-item-link {
-  color: inherit;
-  padding: 1rem;
-  text-decoration: none;
-}
-
-.navbar-list-item-dropdown-toggle {
-  background-color: transparent;
-  border-color: transparent;
-  color: #FFF;
-  display: block;
-  padding: 1rem;
-  text-decoration: none;
-  white-space: nowrap;
-}
-
-.navbar-list-item-dropdown-toggle:hover {
-  cursor: pointer;
-}
-
-  .navbar-list-item-dropdown-toggle::after {
-    border-left: 5px solid transparent;
-    border-right: 5px solid transparent;
-    border-top: 5px dashed;
-    border-top: 5px solid white;
-    content: '';
-    display: inline-block;
-    height: 0;
-    margin-left: 0.5rem;
-    vertical-align: middle;
-    width: 0;
-  }
-
-  .navbar-list-item-dropdown-toggle[aria-expanded="true"] {
-    color: var(--color-orange);
-  }
-  .navbar-list-item-dropdown-toggle[aria-expanded="true"]:after {
-    border-top-color: var(--color-orange);
-  }
-
-
-  .navbar-dropdown-list {
-    background: #FFFFFF;
-    box-shadow: 5px 5px 5px 0 rgba(50, 50, 50, 0.2);
-    padding: 0;
-  }
-
-  li.separator {
-    margin: 0;
-    height: 0;
-    list-style: none;
-    /* TODO fix this */
-    border-bottom: 1px solid black;
-  }
-
-  .navbar-dropdown-list-item {
-    background: #FFF;
-    display: block;
-    list-style: none;
-    white-space: nowrap;
-  }
-
-  a.navbar-dropdown-list-item-link {
-    color: var(--color-orange);
-    display: block;
-    padding: 1rem;
-    text-decoration: none;
-  }
-
-  a.navbar-dropdown-list-item-link:hover,
-  a.navbar-dropdown-list-item-link:focus {
-    background: var(--color-orange);
-    color: var(--color-light);
-  }
-
-
-  /* TODO This rule doesn't make sense so I'm commenting it out for now
-ul[role="menubar"][aria-expanded="true"] li {
-  background: blue;
-} */
-
-  @media (min-width: 992px) {
-    nav.es-navbar {
-      display: flex;
-      flex-direction: row;
-      padding: 0;
-    }
-
-    .navbar-brand {
-      padding: 0;
-      /* TODO using order is not ideal for accessibility */
-      order: 1;
-    }
-
-    .navbar-list {
-      /* TODO using order is not ideal for accessibility */
-      order: 2;
-      align-items: center;
-      display: flex;
-      margin: 0 1rem;
-      padding: 0;
-    }
-
-    .navbar-toggler {
-      display: none; /* TODO remove this properly or else screen readers will still pick up on it */
-    }
-
-    .navbar-dropdown-list {
-      background: #FFFFFF;
-      box-shadow: 5px 5px 5px 0 rgba(50, 50, 50, 0.2);
-      padding: 0;
-      position: absolute;
-      z-index: 100;
-    }
-  }
