+++ addon/templates/components/es-header/navbar-link.hbs
@@ -1,0 +1,1 @@
+{{yield}}
