--- addon/templates/components/es-navbar.hbs
@@ -1,20 +1,0 @@
-<nav class="es-navbar">
-  <a class="navbar-brand-wrapper" href={{navHome}}>
-    <img
-      class="navbar-brand"
-      src="/images/ember-logo-white.svg"
-      alt="home"
-    >
-  </a>
-
-  <button class="navbar-toggler ember-view" onclick={{action this.toggleMenu}} aria-label="Site Navigation" aria-expanded="false">
-    <span class="navbar-toggler-icon"></span>
-  </button>
-
-  <ul class="navbar-list">
-    {{#each navLinks as |link index|}}
-      {{es-navbar/link link=link index=index}}
-    {{/each}}
-  </ul>
-
-</nav>
