+++ app/components/es-header/navbar-link.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-styleguide/components/es-header/navbar-link';
