--- app/components/es-navbar.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-styleguide/components/es-navbar';
