--- docs/components/nav.md
@@ -1,3 +1,0 @@
-# Nav
-
-??
